refactor(mount): Separate mount info functions
This commit separates the mount info related functions into a new file
to improve the maintainability and organization of the code.
